# src/Core/Application/DTOs/NextActions/CreateNextActionDto.cs
namespace Numployable.Application.DTOs.NextActions;

using Common;

public class CreateNextActionDto : BaseDto
{
    public NextActionType? NextActionType {get; set; }

    public string? ActionNote { get; set; }

    public DateTime? ActionDate { get; set; }
}