# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Numployable Api",
    "version": "v1"
  },
  "paths": {
    "/api/JobApplication": {
      "get": {
        "tags": [
          "JobApplication"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JobApplicationDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "JobApplication"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJobApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/JobApplication/{id}": {
      "get": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobApplicationDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobApplicationDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobApplicationDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateJobApplicationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobApplication/expire/{id}": {
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobApplication/processupdate/{id}/{processStatus}": {
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "processStatus",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "$ref": "#/components/schemas/ProcessStatus"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/JobApplication/reject/{id}": {
      "put": {
        "tags": [
          "JobApplication"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/NextAction": {
      "get": {
        "tags": [
          "NextAction"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NextActionDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NextActionDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NextActionDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "NextAction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNextActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNextActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNextActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BaseCommandResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/NextAction/{id}": {
      "get": {
        "tags": [
          "NextAction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NextActionDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "NextAction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNextActionDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNextActionDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNextActionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BaseCommandResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Commute": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "CreateJobApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "roleType": {
            "$ref": "#/components/schemas/RoleType"
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "applicationStatus": {
            "$ref": "#/components/schemas/Status"
          },
          "applicationProcessStatus": {
            "$ref": "#/components/schemas/ProcessStatus"
          },
          "applicationSource": {
            "$ref": "#/components/schemas/Source"
          },
          "advertisedSalary": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "commute": {
            "$ref": "#/components/schemas/Commute"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreateNextActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nextActionType": {
            "$ref": "#/components/schemas/NextActionType"
          },
          "actionNote": {
            "type": "string",
            "nullable": true
          },
          "actionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "JobApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "roleType": {
            "$ref": "#/components/schemas/RoleType"
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "applicationStatus": {
            "$ref": "#/components/schemas/Status"
          },
          "applicationProcessStatus": {
            "$ref": "#/components/schemas/ProcessStatus"
          },
          "applicationSource": {
            "$ref": "#/components/schemas/Source"
          },
          "advertisedSalary": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "commute": {
            "$ref": "#/components/schemas/Commute"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "nextActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NextActionDto"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NextActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nextActionType": {
            "$ref": "#/components/schemas/NextActionType"
          },
          "actionNote": {
            "type": "string",
            "nullable": true
          },
          "actionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "NextActionType": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "ProcessStatus": {
        "enum": [
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
      "RoleType": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "Source": {
        "enum": [
          1,
          2,
          3,
          4
        ],
        "type": "integer",
        "format": "int32"
      },
      "Status": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
      "UpdateJobApplicationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "roleName": {
            "type": "string",
            "nullable": true
          },
          "companyName": {
            "type": "string",
            "nullable": true
          },
          "roleType": {
            "$ref": "#/components/schemas/RoleType"
          },
          "applicationDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "applicationStatus": {
            "$ref": "#/components/schemas/Status"
          },
          "applicationProcessStatus": {
            "$ref": "#/components/schemas/ProcessStatus"
          },
          "applicationSource": {
            "$ref": "#/components/schemas/Source"
          },
          "advertisedSalary": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "commute": {
            "$ref": "#/components/schemas/Commute"
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UpdateNextActionDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "nextActionType": {
            "$ref": "#/components/schemas/NextActionType"
          },
          "actionNote": {
            "type": "string",
            "nullable": true
          },
          "actionDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. \n                      Enter 'Bearer' [space] and then your token in the text input below.\n                      Example: 'Bearer 12345abcdef'",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}